requirements.txt
alembic==1.15.2
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
braintree==4.36.0
bs4==0.0.2
cachelib==0.13.0
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dateparser==1.2.1
debugpy==1.8.12
decorator==5.2.0
Deprecated==1.2.18
Django==5.2
django-allauth==65.9.0
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-countries==7.6.1
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-js-asset==3.1.2
django-model-utils==5.0.0
django-phonenumber-field==8.1.0
django-redis==6.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
email_validator==2.2.0
environ==1.0
executing==2.2.0
Flask==3.1.1
Flask-Caching==2.3.1
flask-cors==5.0.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.0
future==1.0.0
greenlet==3.2.0
gunicorn==20.0.4
h11==0.16.0
hyperopt==0.2.7
idna==3.10
invoke==2.2.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kombu==5.5.4
limits==5.0.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mock==5.2.0
moment==0.12.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
paramiko==4.0.0
parso==0.8.4
passlib==1.7.4
paypalrestsdk==1.13.3
pbr==6.1.1
phonenumbers==9.0.8
pillow==11.2.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
py4j==0.10.9.9
pycparser==2.22
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==0.12.0
python-engineio==4.12.1
python-slugify==8.0.4
python-socketio==5.13.0
pytz==2025.2
pywin32==308
pyzmq==26.2.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.4.0
simple-websocket==1.1.0
six==1.17.0
socks==0
soupsieve==2.7
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
stripe==12.0.0
text-unidecode==1.3
threadpoolctl==3.6.0
times==0.7
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
